# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/album/get/{userId}": {
            "get": {
                "summary": "Retrieves all favorite albums for a specific user",
                "operationId": "201fe8c1db06c7d2be2926b7a48bfa19",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "The user ID to retrieve favorite albums for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of favorite albums",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "albums": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "ID of the album",
                                                        "type": "integer"
                                                    },
                                                    "user_id": {
                                                        "description": "ID of the user who owns the album",
                                                        "type": "integer"
                                                    },
                                                    "title": {
                                                        "description": "Title of the album",
                                                        "type": "string"
                                                    },
                                                    "favorites_count": {
                                                        "description": "Number of favorites in the album",
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No albums found for the user"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/album/create": {
            "post": {
                "summary": "Creates a new favorite album",
                "operationId": "14c2bf57d46885014794cc530037f2ee",
                "requestBody": {
                    "description": "Data needed to create a new album",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "userId"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of the new album",
                                        "type": "string"
                                    },
                                    "userId": {
                                        "description": "User ID of the album owner",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Album created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        },
                                        "album": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID of the newly created album",
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "description": "Title of the album",
                                                    "type": "string"
                                                },
                                                "user_id": {
                                                    "description": "User ID of the album owner",
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "description": "Time when the album was created",
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "description": "Time when the album was last updated",
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input, missing title or userId"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/courses/save": {
            "post": {
                "summary": "Saves a course to a user's favorite album",
                "operationId": "d07ddd2a6c3e2095ff911bab1d7c6c55",
                "requestBody": {
                    "description": "Data needed to save a course to an album",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "courseId",
                                    "albumId"
                                ],
                                "properties": {
                                    "courseId": {
                                        "description": "ID of the course to be saved",
                                        "type": "integer"
                                    },
                                    "albumId": {
                                        "description": "ID of the album where the course is to be saved",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message indicating the course has been saved",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/album/saved/get/{albumId}": {
            "get": {
                "summary": "Retrieves all courses saved in a specified favorite album",
                "operationId": "53cd696412fd38a685f2b453e6627d1c",
                "parameters": [
                    {
                        "name": "albumId",
                        "in": "path",
                        "description": "The ID of the favorite album to retrieve courses from",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful retrieval of saved courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "course_id": {
                                                "description": "ID of the course",
                                                "type": "integer"
                                            },
                                            "course_title": {
                                                "description": "Title of the course",
                                                "type": "string"
                                            },
                                            "course_tags": {
                                                "description": "Tags associated with the course",
                                                "type": "string"
                                            },
                                            "course_thumbnail": {
                                                "description": "URL of the course thumbnail image",
                                                "type": "string"
                                            },
                                            "author": {
                                                "description": "Name of the course author",
                                                "type": "string"
                                            },
                                            "category_name": {
                                                "description": "Name of the category to which the course belongs",
                                                "type": "string"
                                            },
                                            "course_price": {
                                                "description": "Price of the course",
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "number_of_lessons": {
                                                "description": "Total number of lessons in the course",
                                                "type": "integer"
                                            },
                                            "number_of_enrollments": {
                                                "description": "Total number of enrollments in the course",
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No saved courses in this album",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "description": "Error message indicating no courses found in the album",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}